refactor: combine host-descriptiopn with room-description
