+++ project/src/components/room-reviews/room-reviews.tsx
@@ -12,7 +12,7 @@ type RoomReviewProps = {
 
 function RoomReviews({ isUserLoggedIn, offerId, reviews }: RoomReviewProps): JSX.Element {
   return (
-    <>
+    <section className="property__reviews reviews">
       <h2 className="reviews__title">
         Reviews &middot; <span className="reviews__amount">{reviews.length}</span>
       </h2>
@@ -41,7 +41,7 @@ function RoomReviews({ isUserLoggedIn, offerId, reviews }: RoomReviewProps): JSX
       }
 
       { isUserLoggedIn && <RoomReviewForm key={offerId} /> }
-    </>
+    </section>
   );
 }
