+++ project/src/pages/room/room.tsx
@@ -5,16 +5,15 @@ import { useOfferReviews } from 'src/hooks/use-offer-reviews';
 import { useFoundOffer } from 'src/hooks/use-found-offer';
 import { getAuthorizationStatus } from 'src/store/user/user.selectors';
 import OfferCards from 'src/components/offer-сards/offer-cards';
-import RoomDescription from 'src/components/room-description/room-description';
 import RoomGallery from 'src/components/room-gallery/room-gallery';
 import RoomHardwareFeatures from 'src/components/room-hardware-features/room-hardware-features';
-import RoomHost from 'src/components/room-host/room-host';
 import RoomReviews from 'src/components/room-reviews/room-reviews';
 import GeoMap from 'src/components/geo-map/geo-map';
 import { Spinner } from 'src/components/spinner/spinner';
 import { getPercentFromRating, capitalizeFirstLetter } from 'src/utils/utils';
 import { AppRoute, NEARBY_OFFERS_LIMIT_COUNT } from 'src/consts/consts';
 import { AuthorizationStatus } from 'src/consts/api';
+import RoomHostDescription from 'src/components/room-host-description/room-host-description';
 
 type RoomProps = {
   headerBlock?: JSX.Element;
@@ -34,8 +33,7 @@ function Room({ headerBlock }: RoomProps): JSX.Element {
   if (!offer) {
     return <Spinner text={'Loading offer ...'} />;
   }
-  // eslint-disable-next-line no-console
-  console.log(offerId, authorizationStatus, offer, nearbyOffers, reviews);
+
   return (
     <div className="page">
       {headerBlock}
@@ -79,19 +77,15 @@ function Room({ headerBlock }: RoomProps): JSX.Element {
                 <span className="property__price-text">&nbsp;night</span>
               </div>
               <RoomHardwareFeatures goods={offer.goods}/>
-              {/*  TODO move whole following block to <Host>? */}
-              <div className="property__host">
-                <h2 className="property__host-title">Meet the host</h2>
-                <RoomHost host={offer.host} />
-                <RoomDescription description={offer.description} />
-              </div>
-              <section className="property__reviews reviews">
-                <RoomReviews
-                  isUserLoggedIn={authorizationStatus === AuthorizationStatus.Authorized}
-                  offerId={offer.id}
-                  reviews={reviews}
-                />
-              </section>
+              <RoomHostDescription
+                host={offer.host}
+                description={offer.description}
+              />
+              <RoomReviews
+                isUserLoggedIn={authorizationStatus === AuthorizationStatus.Authorized}
+                offerId={offer.id}
+                reviews={reviews}
+              />
             </div>
           </div>
           <GeoMap
